refactor: move calcs to yearly.py
