+++ tests/test_wmrc.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pandas as pd
+
 from wmrc import main
